Extract single token and multiple token useSele- ction into custom useSelection hook
