+++ src/pages/textAnalyzer/TextAnalyzer.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 import { getEventListeners } from "events";
+import useSelection from "./utils/useSelection";
 
 const TextAnalyzer = () => {
 
@@ -14,6 +15,8 @@ const TextAnalyzer = () => {
   let lastSelectedTokenRef = useRef(null)
   let IscombiningModeEngaged = useRef(false)
   let isCTRLPressed = useRef(false)
+
+  useSelection(currentModeRef);
 
   useEffect(() => {
     currentModeRef.current = currentMode;
@@ -42,148 +45,6 @@ const TextAnalyzer = () => {
       document.removeEventListener('selectionchange', handleSelectionChange);
     };
   }, []);
-
-  useEffect(() => {
-
-    const selection = window.getSelection();
-
-    // const handleSelectionAndMultiSelect = (e) => {
-
-      const handleSelection = (e) => {
-      // NOTE: Remember to use currentModeRef.current instead of state variable currentMode
-
-        if (!selection || selection.rangeCount === 0) return;
-
-        // add or remove selection class
-
-          const range = selection.getRangeAt(0);
-          const spanElement = range.commonAncestorContainer.parentElement as HTMLElement;
-
-          const isSpanElementToken = spanElement.classList.contains('token');
-          const isSpanElementSelected = spanElement.classList.contains('selected');
-
-
-          if ((spanElement.tagName) === 'SPAN' && isSpanElementToken) {
-
-            if (!isSpanElementSelected) {
-              spanElement.classList.add('selected');
-
-              if (spanElement.classList.contains("hovering")) {
-                spanElement.classList.remove("hovering");
-              };
-  
-              const selectionList = document.getElementsByClassName("selected");
-  
-              if (currentModeRef.current !== 'editTokenList') {
-                for (let i = 0; i < selectionList.length; i++) {
-                  if (selectionList[i] !== spanElement)
-                  selectionList[i].classList.remove("selected");
-                };
-              }
-            }
-
-          }
-      }
-
-      type SelectionDirection = 'backward' | 'forward'
-
-      const getSelectionDirection = () => {
-        // dones't include cases where selection is within same node. 
-      
-        const sel = window.getSelection();
-
-        if (!sel) {
-          return
-        };
-
-        if (sel.rangeCount === 0) {
-          return
-        };
-
-        const anchorNode = sel.anchorNode;
-        const focusNode = sel.focusNode;
-
-        const position = anchorNode.compareDocumentPosition(focusNode);
-
-        if (!position && sel.anchorOffset < sel.focusOffset || position === Node.DOCUMENT_POSITION_PRECEDING) {
-          return 'backward'
-        } else {
-          return 'forward'
-        }
-      }
-
-      const handleMultiSelect = (e) => {
-        const sel = document.getSelection();
-
-        if (!sel || !sel.getRangeAt(0)) {
-          return;
-        }
-
-        const anchorNode = sel.anchorNode.parentElement as Element;
-        const focusNode = sel.focusNode.parentElement as Element;
-
-        
-        if (
-          anchorNode === focusNode ||
-          !anchorNode ||
-          !focusNode ||
-          !anchorNode.classList ||
-          !focusNode.classList ||
-          (!anchorNode.classList.contains("token") ||
-          !focusNode.classList.contains("token"))
-          ) {
-            return;
-          }
-
-
-      
-          anchorNode.classList.add('selected');
-
-          const selectNextSibling = (currentSpan: Element, finalSpan: Element) => {
-            currentSpan.classList.add('selected');
-  
-            const nextSibbling: Element = currentSpan.nextElementSibling;
-  
-            if (nextSibbling !== finalSpan) {
-              nextSibbling.classList.add('selected');
-              selectNextSibling(nextSibbling, finalSpan);
-            } else {
-              finalSpan.classList.add('selected');
-            }
-          }
-  
-          const selectPrevSibling = (currentSpan: Element, finalSpan: Element) => {
-            currentSpan.classList.add('selected');
-  
-            const prevSibbling: Element = currentSpan.nextElementSibling;
-  
-            if (prevSibbling !== finalSpan) {
-              prevSibbling.classList.add('selected');
-              selectPrevSibling(prevSibbling, finalSpan);
-            } else {
-              finalSpan.classList.add('selected');
-            }
-          }
-  
-          const Direction = getSelectionDirection();
-  
-          if (Direction === 'forward')
-            selectNextSibling(anchorNode, focusNode);
-          else {
-            selectPrevSibling(focusNode, anchorNode);
-          }
-
-          sel.removeAllRanges();
-    }
-
-    document.addEventListener('selectionchange', handleSelection);
-    document.addEventListener('mouseup', handleMultiSelect);
-
-    return () => {
-      document.removeEventListener('selectionchange', handleSelection);
-      document.removeEventListener('mouseup', handleMultiSelect);
-    }
-  }, [])
 
   useEffect(() => {
